facebook-openai-chatbot/requirements.txt
flask==2.2.3
python-dotenv==1.0.0
requests==2.28.2
openai==0.28.1
gspread==5.9.0
oauth2client==4.1.3
python-facebook-api==0.15.0
gunicorn==20.1.0
